+++ package.json
@@ -13,11 +13,11 @@
   "license": "ISC",
   "devDependencies": {
     "@types/dotenv": "^8.2.0",
-    "@types/typescript": "^2.0.0",
     "@typescript-eslint/eslint-plugin": "^6.2.0",
     "@typescript-eslint/parser": "^6.2.0",
     "cors": "^2.8.5",
     "csv-parse": "^5.4.0",
+    "express": "^4.18.2",
     "helmet": "^7.0.0",
     "nodemon": "^3.0.1"
   },
